fix naming mistake
Fix suggested by Ferenc Dobi

If you are trying to put a check constraint on the 4th column, as
other tests use the same input for it ("a") then it is impossible to
pass.

On the other hand, you can put a constraint on the third column, as
other tests always insert "Europe" into it, while this one does "a".